requirements.txt
asgiref==3.5.2
Django==4.1.2
django-admin-honeypot==1.1.0
django-ipware==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
